BooksReseñas/requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
attrs==23.2.0
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==42.0.8
defusedxml==0.8.0rc2
Django==5.0.6
django-captcha-admin==0.2
django-crum==0.7.9
django-mssql-backend==2.8.1
django-ranged-response==0.2.0
django-recaptcha==4.0.0
django-simple-captcha==0.6.0
frozenlist==1.4.1
google-api-core==2.19.0
google-auth==2.30.0
google-cloud-recaptcha-enterprise==1.20.0
googleapis-common-protos==1.63.1
grpcio==1.64.1
grpcio-status==1.62.2
idna==3.7
mssql-django==1.5
multidict==6.0.5
oauthlib==3.2.2
pillow==10.3.0
polib==1.2.0
proto-plus==1.24.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyodbc==5.1.0
pypng==0.20220715.0
python-decouple==3.8
python3-openid==3.2.0
pytz==2024.1
qrcode==7.4.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
yarl==1.9.4
